ENG-3704 Amend description messages, rename FileHelper to FSService
