+++ src/services/fs-service.ts
@@ -5,6 +5,7 @@ import debugFactory from './debug-factory-service'
 import { RESOURCES_FOLDER } from '../paths'
 
 const ALLOWED_BUNDLE_NAME_REGEXP = /^[\w-]+$/
+const JSON_INDENTATION_SPACES = 4
 
 export default class FSService {
   private static debug = debugFactory(FSService)
@@ -64,5 +65,12 @@ export default class FSService {
     if (!fs.existsSync(this.getBundleFilePath(...subDirectories))) {
       fs.mkdirSync(this.getBundleFilePath(...subDirectories))
     }
+  }
+
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  public static writeJSON(filePath: string, data: any): void {
+    fs.writeFileSync(
+      filePath,
+      JSON.stringify(data, null, JSON_INDENTATION_SPACES))
   }
 }
